+++ espflash/Cargo.toml
@@ -55,7 +55,7 @@ strum = { version = "0.26.1", features = ["derive"] }
 thiserror = "1.0.56"
 toml = { version = "0.8.10", optional = true }
 update-informer = { version = "1.1.0", optional = true }
-xmas-elf        = { version = "0.9.1", optional = true }
+xmas-elf        = { version = "0.9.1" }
 
 [target.'cfg(unix)'.dependencies]
 libc = "0.2.153"
@@ -87,4 +87,4 @@ cli = [
 serialport = ["flashing", "dep:serialport", "dep:slip-codec", "dep:regex"]
 
 # enables flash stubs and stub commands
-flashing = ["xmas-elf", "dep:toml"]
+flashing = ["dep:toml"]
