+++ espflash/src/targets/esp32h2.rs
@@ -5,7 +5,7 @@ use crate::connection::Connection;
 use crate::{
     elf::FirmwareImage,
     error::Error,
-    flasher::{FlashData, FlashFrequency},
+    flash_data::{FlashData, FlashFrequency},
     image_format::IdfBootloaderFormat,
     targets::{Chip, Esp32Params, ReadEFuse, SpiRegisters, Target, XtalFrequency},
 };
