+++ espflash/src/targets/flash_target/mod.rs
@@ -1,8 +1,7 @@
 pub(crate) use self::ram::MAX_RAM_BLOCK_SIZE;
 pub use self::{esp32::Esp32Target, ram::RamTarget};
 #[cfg(feature = "serialport")]
-use crate::connection::Connection;
-use crate::{elf::RomSegment, error::Error, flasher::ProgressCallbacks};
+use crate::{connection::Connection, elf::RomSegment, error::Error, progress::ProgressCallbacks};
 
 mod esp32;
 mod ram;
